update: rename util
