+++ app/src/main/java/com/byl/mvvm/ui/base/BaseActivity.kt
@@ -15,7 +15,7 @@ import com.byl.mvvm.event.EventMessage
 import com.byl.mvvm.ext.toast
 import com.byl.mvvm.ui.dialog.LoadingDialog
 import com.byl.mvvm.util.Logg
-import com.byl.mvvm.utils.GenericParadigmUtil
+import com.byl.mvvm.util.GenericParadigmUtils
 import org.greenrobot.eventbus.Subscribe
 
 abstract class BaseActivity<VM : BaseViewModel<VB>, VB : ViewBinding> : AppCompatActivity(), IView {
@@ -30,14 +30,14 @@ abstract class BaseActivity<VM : BaseViewModel<VB>, VB : ViewBinding> : AppCompa
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         initResources()
-        var pathfinders = ArrayList<GenericParadigmUtil.Pathfinder>()
-        pathfinders.add(GenericParadigmUtil.Pathfinder(0, 0))
-        val clazzVM = GenericParadigmUtil.parseGenericParadigm(javaClass, pathfinders) as Class<VM>
+        var pathfinders = ArrayList<GenericParadigmUtils.Pathfinder>()
+        pathfinders.add(GenericParadigmUtils.Pathfinder(0, 0))
+        val clazzVM = GenericParadigmUtils.parseGenericParadigm(javaClass, pathfinders) as Class<VM>
         vm = ViewModelProvider(this).get(clazzVM)
 
         pathfinders = ArrayList()
-        pathfinders.add(GenericParadigmUtil.Pathfinder(0, 1))
-        val clazzVB = GenericParadigmUtil.parseGenericParadigm(javaClass, pathfinders)
+        pathfinders.add(GenericParadigmUtils.Pathfinder(0, 1))
+        val clazzVB = GenericParadigmUtils.parseGenericParadigm(javaClass, pathfinders)
         val method = clazzVB.getMethod("inflate", LayoutInflater::class.java)
         vb = method.invoke(null, layoutInflater) as VB
